requirements.txt
pandas
numpy
requests
networkx
matplotlib
scipy
gravis
ipython
dash
openai
python-dotenv